solana_alpha_hunter/requirements.txt
streamlit==1.35.0
streamlit-autorefresh==1.0.1
lightgbm==4.3.0
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
networkx==3.3
python-louvain==0.16
plotly==5.22.0
apscheduler==3.10.4
python-telegram-bot==21.3
websockets==12.0
aiohttp==3.9.5
httpx==0.27.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
rich==13.7.1
tabulate==0.9.0
